removed threading
