--- changelog/PORT-5475.feature.md
@@ -1,1 +1,0 @@
-Added handling for kafka consumer empty partition assignment and shutting the application down with an error
