--- changelog/PORT-5475.improvement.md
@@ -1,1 +1,0 @@
-Changed the Kafka consumer to run in the event loop in async instead of sync in another thread
