--- changelog/PORT-5475_2.feature.md
@@ -1,1 +1,0 @@
-Added QOL decorator to help with caching the third party response
